Clear lib.rs file
